[FIX]: 엔티티 변경으로 인한  전체적인 수정

+++ backend/src/main/java/com/school/chick/domain/repository/AuthRefreshSaveRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface AuthRefreshSaveRepository extends JpaRepository<AuthRefreshSave, Long> {
+public interface AuthRefreshSaveRepository extends JpaRepository<AuthRefreshSave, Integer> {
     AuthRefreshSave findByRefreshToken(String refreshToken);
 }
